requirements.txt
greenlet==3.2.4
packaging==25.0
pillow==11.3.0
playwright==1.54.0
pyee==13.0.0
pytesseract==0.3.13
python-dotenv==1.1.1
slack_sdk==3.36.0
typing_extensions==4.14.1
